Introduce Core Data to store completed transactions for easy sort/search.
Upon app launch, "transactions.db" would be opened for Core Data storage
of the completed transactions. If it doesn't exist yet, it will be created
and populated from "history.plist". In either case, NSManagedObjectContext
is obtained and assigned to "context" property of the main table view
controller for pending transactions and later to the other table view
controller for completed transactions. Therefore, the context is shared
between the two table views and no synchronization is needed.

The history view is a new table view backed by NSFetchedResultsController.
CoreDataTableViewController is adopted from Stanford CS193p class with two
modifications:
- NSLog is commented out when performFetch has error.
- Release the old FetchedResultsController when a new one is set. It may
not be needed when ARC is enabled. Until then, it is important to release
the old one to unregister the delegate.

The main table view doesn't render completed transactions by "historyView"
flag any more. However, completed transactions are still saved to "history.
plist" file as well as the Core Data storage. "history.plist" file is still
kept so that it can be used to populate an external DB as it is a plain
text file.